+++ next/src/components/ui/icons/ErrorIcon.test.tsx
@@ -1,0 +1,8 @@
+import { render } from '@testing-library/react'
+
+import ErrorIcon from './ErrorIcon'
+
+it('ErrorIcon renders correctly', () => {
+  const { asFragment } = render(<ErrorIcon />)
+  expect(asFragment()).toMatchSnapshot()
+})