+++ next/src/routeTree.gen.ts
@@ -32,6 +32,9 @@ import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationPublic
 import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationIndexRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/personalization/index'
 import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationNewRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/personalization/new'
 import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationRouteRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/personalization/route'
+import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQReleasesIndexRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/releases/index'
+import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQReleasesNewRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/releases/new'
+import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQReleasesRouteRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/releases/route'
 import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQTcmCompositionRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/tcm-composition'
 import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQVariablesIndexRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/variables/index'
 import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQVariablesNewRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/variables/new'
@@ -165,6 +168,12 @@ const LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationRouteRoute =
       getParentRoute: () => LayoutQuestionnaireQuestionnaireIdLayoutQRoute,
     } as any,
   )
+const LayoutQuestionnaireQuestionnaireIdLayoutQReleasesRouteRoute =
+  LayoutQuestionnaireQuestionnaireIdLayoutQReleasesRouteRouteImport.update({
+    id: '/releases',
+    path: '/releases',
+    getParentRoute: () => LayoutQuestionnaireQuestionnaireIdLayoutQRoute,
+  } as any)
 const LayoutQuestionnaireQuestionnaireIdLayoutQTcmCompositionRoute =
   LayoutQuestionnaireQuestionnaireIdLayoutQTcmCompositionRouteImport.update({
     id: '/tcm-composition',
@@ -283,6 +292,20 @@ const LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationNewRoute =
         LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationRouteRoute,
     } as any,
   )
+const LayoutQuestionnaireQuestionnaireIdLayoutQReleasesIndexRoute =
+  LayoutQuestionnaireQuestionnaireIdLayoutQReleasesIndexRouteImport.update({
+    id: '/',
+    path: '/',
+    getParentRoute: () =>
+      LayoutQuestionnaireQuestionnaireIdLayoutQReleasesRouteRoute,
+  } as any)
+const LayoutQuestionnaireQuestionnaireIdLayoutQReleasesNewRoute =
+  LayoutQuestionnaireQuestionnaireIdLayoutQReleasesNewRouteImport.update({
+    id: '/new',
+    path: '/new',
+    getParentRoute: () =>
+      LayoutQuestionnaireQuestionnaireIdLayoutQReleasesRouteRoute,
+  } as any)
 const LayoutQuestionnaireQuestionnaireIdLayoutQVariablesIndexRoute =
   LayoutQuestionnaireQuestionnaireIdLayoutQVariablesIndexRouteImport.update({
     id: '/',
@@ -392,6 +415,7 @@ export interface FileRoutesByFullPath {
   '/questionnaire/$questionnaireId/multimode': typeof LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeRouteRouteWithChildren
   '/questionnaire/$questionnaireId/nomenclatures': typeof LayoutQuestionnaireQuestionnaireIdLayoutQNomenclaturesRouteRouteWithChildren
   '/questionnaire/$questionnaireId/personalization': typeof LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationRouteRouteWithChildren
+  '/questionnaire/$questionnaireId/releases': typeof LayoutQuestionnaireQuestionnaireIdLayoutQReleasesRouteRouteWithChildren
   '/questionnaire/$questionnaireId/variables': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVariablesRouteRouteWithChildren
   '/questionnaire/$questionnaireId/composition': typeof LayoutQuestionnaireQuestionnaireIdLayoutQCompositionRoute
   '/questionnaire/$questionnaireId/duplicate-variables': typeof LayoutQuestionnaireQuestionnaireIdLayoutQDuplicateVariablesRoute
@@ -407,6 +431,7 @@ export interface FileRoutesByFullPath {
   '/questionnaire/$questionnaireId/multimode/edit': typeof LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeEditRoute
   '/questionnaire/$questionnaireId/personalization/$publicEnemyId': typeof LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationPublicEnemyIdRoute
   '/questionnaire/$questionnaireId/personalization/new': typeof LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationNewRoute
+  '/questionnaire/$questionnaireId/releases/new': typeof LayoutQuestionnaireQuestionnaireIdLayoutQReleasesNewRoute
   '/questionnaire/$questionnaireId/variables/new': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVariablesNewRoute
   '/questionnaire/$questionnaireId/articulation/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQArticulationIndexRoute
   '/questionnaire/$questionnaireId/codes-lists/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsIndexRoute
@@ -414,6 +439,7 @@ export interface FileRoutesByFullPath {
   '/questionnaire/$questionnaireId/multimode/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeIndexRoute
   '/questionnaire/$questionnaireId/nomenclatures/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQNomenclaturesIndexRoute
   '/questionnaire/$questionnaireId/personalization/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationIndexRoute
+  '/questionnaire/$questionnaireId/releases/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQReleasesIndexRoute
   '/questionnaire/$questionnaireId/variables/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVariablesIndexRoute
   '/questionnaire/$questionnaireId/variables/variable/$variableId': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVariablesVariableVariableIdRoute
   '/questionnaire/$questionnaireId/version/$versionId/articulation': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdArticulationRoute
@@ -443,6 +469,7 @@ export interface FileRoutesByTo {
   '/questionnaire/$questionnaireId/multimode/edit': typeof LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeEditRoute
   '/questionnaire/$questionnaireId/personalization/$publicEnemyId': typeof LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationPublicEnemyIdRoute
   '/questionnaire/$questionnaireId/personalization/new': typeof LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationNewRoute
+  '/questionnaire/$questionnaireId/releases/new': typeof LayoutQuestionnaireQuestionnaireIdLayoutQReleasesNewRoute
   '/questionnaire/$questionnaireId/variables/new': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVariablesNewRoute
   '/questionnaire/$questionnaireId/articulation': typeof LayoutQuestionnaireQuestionnaireIdLayoutQArticulationIndexRoute
   '/questionnaire/$questionnaireId/codes-lists': typeof LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsIndexRoute
@@ -450,6 +477,7 @@ export interface FileRoutesByTo {
   '/questionnaire/$questionnaireId/multimode': typeof LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeIndexRoute
   '/questionnaire/$questionnaireId/nomenclatures': typeof LayoutQuestionnaireQuestionnaireIdLayoutQNomenclaturesIndexRoute
   '/questionnaire/$questionnaireId/personalization': typeof LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationIndexRoute
+  '/questionnaire/$questionnaireId/releases': typeof LayoutQuestionnaireQuestionnaireIdLayoutQReleasesIndexRoute
   '/questionnaire/$questionnaireId/variables': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVariablesIndexRoute
   '/questionnaire/$questionnaireId/variables/variable/$variableId': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVariablesVariableVariableIdRoute
   '/questionnaire/$questionnaireId/version/$versionId/articulation': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdArticulationRoute
@@ -477,6 +505,7 @@ export interface FileRoutesById {
   '/_layout/questionnaire/$questionnaireId/_layout-q/multimode': typeof LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeRouteRouteWithChildren
   '/_layout/questionnaire/$questionnaireId/_layout-q/nomenclatures': typeof LayoutQuestionnaireQuestionnaireIdLayoutQNomenclaturesRouteRouteWithChildren
   '/_layout/questionnaire/$questionnaireId/_layout-q/personalization': typeof LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationRouteRouteWithChildren
+  '/_layout/questionnaire/$questionnaireId/_layout-q/releases': typeof LayoutQuestionnaireQuestionnaireIdLayoutQReleasesRouteRouteWithChildren
   '/_layout/questionnaire/$questionnaireId/_layout-q/variables': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVariablesRouteRouteWithChildren
   '/_layout/questionnaire/$questionnaireId/_layout-q/composition': typeof LayoutQuestionnaireQuestionnaireIdLayoutQCompositionRoute
   '/_layout/questionnaire/$questionnaireId/_layout-q/duplicate-variables': typeof LayoutQuestionnaireQuestionnaireIdLayoutQDuplicateVariablesRoute
@@ -492,6 +521,7 @@ export interface FileRoutesById {
   '/_layout/questionnaire/$questionnaireId/_layout-q/multimode/edit': typeof LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeEditRoute
   '/_layout/questionnaire/$questionnaireId/_layout-q/personalization/$publicEnemyId': typeof LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationPublicEnemyIdRoute
   '/_layout/questionnaire/$questionnaireId/_layout-q/personalization/new': typeof LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationNewRoute
+  '/_layout/questionnaire/$questionnaireId/_layout-q/releases/new': typeof LayoutQuestionnaireQuestionnaireIdLayoutQReleasesNewRoute
   '/_layout/questionnaire/$questionnaireId/_layout-q/variables/new': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVariablesNewRoute
   '/_layout/questionnaire/$questionnaireId/_layout-q/articulation/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQArticulationIndexRoute
   '/_layout/questionnaire/$questionnaireId/_layout-q/codes-lists/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsIndexRoute
@@ -499,6 +529,7 @@ export interface FileRoutesById {
   '/_layout/questionnaire/$questionnaireId/_layout-q/multimode/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeIndexRoute
   '/_layout/questionnaire/$questionnaireId/_layout-q/nomenclatures/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQNomenclaturesIndexRoute
   '/_layout/questionnaire/$questionnaireId/_layout-q/personalization/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationIndexRoute
+  '/_layout/questionnaire/$questionnaireId/_layout-q/releases/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQReleasesIndexRoute
   '/_layout/questionnaire/$questionnaireId/_layout-q/variables/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVariablesIndexRoute
   '/_layout/questionnaire/$questionnaireId/_layout-q/variables/variable/$variableId': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVariablesVariableVariableIdRoute
   '/_layout/questionnaire/$questionnaireId/_layout-q/version/$versionId/articulation': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdArticulationRoute
@@ -525,6 +556,7 @@ export interface FileRouteTypes {
     | '/questionnaire/$questionnaireId/multimode'
     | '/questionnaire/$questionnaireId/nomenclatures'
     | '/questionnaire/$questionnaireId/personalization'
+    | '/questionnaire/$questionnaireId/releases'
     | '/questionnaire/$questionnaireId/variables'
     | '/questionnaire/$questionnaireId/composition'
     | '/questionnaire/$questionnaireId/duplicate-variables'
@@ -540,6 +572,7 @@ export interface FileRouteTypes {
     | '/questionnaire/$questionnaireId/multimode/edit'
     | '/questionnaire/$questionnaireId/personalization/$publicEnemyId'
     | '/questionnaire/$questionnaireId/personalization/new'
+    | '/questionnaire/$questionnaireId/releases/new'
     | '/questionnaire/$questionnaireId/variables/new'
     | '/questionnaire/$questionnaireId/articulation/'
     | '/questionnaire/$questionnaireId/codes-lists/'
@@ -547,6 +580,7 @@ export interface FileRouteTypes {
     | '/questionnaire/$questionnaireId/multimode/'
     | '/questionnaire/$questionnaireId/nomenclatures/'
     | '/questionnaire/$questionnaireId/personalization/'
+    | '/questionnaire/$questionnaireId/releases/'
     | '/questionnaire/$questionnaireId/variables/'
     | '/questionnaire/$questionnaireId/variables/variable/$variableId'
     | '/questionnaire/$questionnaireId/version/$versionId/articulation'
@@ -576,6 +610,7 @@ export interface FileRouteTypes {
     | '/questionnaire/$questionnaireId/multimode/edit'
     | '/questionnaire/$questionnaireId/personalization/$publicEnemyId'
     | '/questionnaire/$questionnaireId/personalization/new'
+    | '/questionnaire/$questionnaireId/releases/new'
     | '/questionnaire/$questionnaireId/variables/new'
     | '/questionnaire/$questionnaireId/articulation'
     | '/questionnaire/$questionnaireId/codes-lists'
@@ -583,6 +618,7 @@ export interface FileRouteTypes {
     | '/questionnaire/$questionnaireId/multimode'
     | '/questionnaire/$questionnaireId/nomenclatures'
     | '/questionnaire/$questionnaireId/personalization'
+    | '/questionnaire/$questionnaireId/releases'
     | '/questionnaire/$questionnaireId/variables'
     | '/questionnaire/$questionnaireId/variables/variable/$variableId'
     | '/questionnaire/$questionnaireId/version/$versionId/articulation'
@@ -609,6 +645,7 @@ export interface FileRouteTypes {
     | '/_layout/questionnaire/$questionnaireId/_layout-q/multimode'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/nomenclatures'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/personalization'
+    | '/_layout/questionnaire/$questionnaireId/_layout-q/releases'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/variables'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/composition'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/duplicate-variables'
@@ -624,6 +661,7 @@ export interface FileRouteTypes {
     | '/_layout/questionnaire/$questionnaireId/_layout-q/multimode/edit'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/personalization/$publicEnemyId'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/personalization/new'
+    | '/_layout/questionnaire/$questionnaireId/_layout-q/releases/new'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/variables/new'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/articulation/'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/codes-lists/'
@@ -631,6 +669,7 @@ export interface FileRouteTypes {
     | '/_layout/questionnaire/$questionnaireId/_layout-q/multimode/'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/nomenclatures/'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/personalization/'
+    | '/_layout/questionnaire/$questionnaireId/_layout-q/releases/'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/variables/'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/variables/variable/$variableId'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/version/$versionId/articulation'
@@ -782,6 +821,13 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationRouteRouteImport
       parentRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQRoute
     }
+    '/_layout/questionnaire/$questionnaireId/_layout-q/releases': {
+      id: '/_layout/questionnaire/$questionnaireId/_layout-q/releases'
+      path: '/releases'
+      fullPath: '/questionnaire/$questionnaireId/releases'
+      preLoaderRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQReleasesRouteRouteImport
+      parentRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQRoute
+    }
     '/_layout/questionnaire/$questionnaireId/_layout-q/tcm-composition': {
       id: '/_layout/questionnaire/$questionnaireId/_layout-q/tcm-composition'
       path: '/tcm-composition'
@@ -893,6 +939,20 @@ declare module '@tanstack/react-router' {
       fullPath: '/questionnaire/$questionnaireId/personalization/new'
       preLoaderRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationNewRouteImport
       parentRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationRouteRoute
+    }
+    '/_layout/questionnaire/$questionnaireId/_layout-q/releases/': {
+      id: '/_layout/questionnaire/$questionnaireId/_layout-q/releases/'
+      path: '/'
+      fullPath: '/questionnaire/$questionnaireId/releases/'
+      preLoaderRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQReleasesIndexRouteImport
+      parentRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQReleasesRouteRoute
+    }
+    '/_layout/questionnaire/$questionnaireId/_layout-q/releases/new': {
+      id: '/_layout/questionnaire/$questionnaireId/_layout-q/releases/new'
+      path: '/new'
+      fullPath: '/questionnaire/$questionnaireId/releases/new'
+      preLoaderRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQReleasesNewRouteImport
+      parentRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQReleasesRouteRoute
     }
     '/_layout/questionnaire/$questionnaireId/_layout-q/variables/': {
       id: '/_layout/questionnaire/$questionnaireId/_layout-q/variables/'
@@ -1098,6 +1158,24 @@ const LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationRouteRouteWithChil
     LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationRouteRouteChildren,
   )
 
+interface LayoutQuestionnaireQuestionnaireIdLayoutQReleasesRouteRouteChildren {
+  LayoutQuestionnaireQuestionnaireIdLayoutQReleasesNewRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQReleasesNewRoute
+  LayoutQuestionnaireQuestionnaireIdLayoutQReleasesIndexRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQReleasesIndexRoute
+}
+
+const LayoutQuestionnaireQuestionnaireIdLayoutQReleasesRouteRouteChildren: LayoutQuestionnaireQuestionnaireIdLayoutQReleasesRouteRouteChildren =
+  {
+    LayoutQuestionnaireQuestionnaireIdLayoutQReleasesNewRoute:
+      LayoutQuestionnaireQuestionnaireIdLayoutQReleasesNewRoute,
+    LayoutQuestionnaireQuestionnaireIdLayoutQReleasesIndexRoute:
+      LayoutQuestionnaireQuestionnaireIdLayoutQReleasesIndexRoute,
+  }
+
+const LayoutQuestionnaireQuestionnaireIdLayoutQReleasesRouteRouteWithChildren =
+  LayoutQuestionnaireQuestionnaireIdLayoutQReleasesRouteRoute._addFileChildren(
+    LayoutQuestionnaireQuestionnaireIdLayoutQReleasesRouteRouteChildren,
+  )
+
 interface LayoutQuestionnaireQuestionnaireIdLayoutQVariablesRouteRouteChildren {
   LayoutQuestionnaireQuestionnaireIdLayoutQVariablesNewRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQVariablesNewRoute
   LayoutQuestionnaireQuestionnaireIdLayoutQVariablesIndexRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQVariablesIndexRoute
@@ -1159,6 +1237,7 @@ interface LayoutQuestionnaireQuestionnaireIdLayoutQRouteChildren {
   LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeRouteRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeRouteRouteWithChildren
   LayoutQuestionnaireQuestionnaireIdLayoutQNomenclaturesRouteRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQNomenclaturesRouteRouteWithChildren
   LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationRouteRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationRouteRouteWithChildren
+  LayoutQuestionnaireQuestionnaireIdLayoutQReleasesRouteRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQReleasesRouteRouteWithChildren
   LayoutQuestionnaireQuestionnaireIdLayoutQVariablesRouteRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQVariablesRouteRouteWithChildren
   LayoutQuestionnaireQuestionnaireIdLayoutQCompositionRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQCompositionRoute
   LayoutQuestionnaireQuestionnaireIdLayoutQDuplicateVariablesRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQDuplicateVariablesRoute
@@ -1184,6 +1263,8 @@ const LayoutQuestionnaireQuestionnaireIdLayoutQRouteChildren: LayoutQuestionnair
       LayoutQuestionnaireQuestionnaireIdLayoutQNomenclaturesRouteRouteWithChildren,
     LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationRouteRoute:
       LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationRouteRouteWithChildren,
+    LayoutQuestionnaireQuestionnaireIdLayoutQReleasesRouteRoute:
+      LayoutQuestionnaireQuestionnaireIdLayoutQReleasesRouteRouteWithChildren,
     LayoutQuestionnaireQuestionnaireIdLayoutQVariablesRouteRoute:
       LayoutQuestionnaireQuestionnaireIdLayoutQVariablesRouteRouteWithChildren,
     LayoutQuestionnaireQuestionnaireIdLayoutQCompositionRoute: